Rip out lots of weird stuff and fix bugs
We reviewed SIMP-2893 earlier in the week and realized that most of the
weirder logic (and a few subtle bugs) could be drastically reduced by
breaking the yum repos into lighter individual profiles and just
including them by through classification instead of overwhelming
``yum::simp``.

This result is more flexible and easier to grok than the
previous method, and it takes _much_ less work to maintain.

I've traded the magic ruby mangling functions out for simpler
puppet language functions.  They could be simpler still (or gone), but
we're still not sure whether some of the weird options from the legacy
code is actually needed by anyone.